+++ FishingAssistant/ModEntry.cs
@@ -19,6 +19,7 @@ internal sealed class ModEntry : Mod
     private StarterFishingRodService? starterFishingRod;
     private DebugWarpService? debugWarp;
     private BaitAttachmentService? baitAttachment;
+    private TackleAttachmentService? tackleAttachment;
 
     public override void Entry(IModHelper helper)
     {
@@ -31,6 +32,7 @@ internal sealed class ModEntry : Mod
         this.starterFishingRod = new StarterFishingRodService(this.Monitor, key => helper.Translation.Get(key));
         this.debugWarp = new DebugWarpService(this.Monitor, key => helper.Translation.Get(key));
         this.baitAttachment = new BaitAttachmentService(this.Monitor, key => helper.Translation.Get(key));
+        this.tackleAttachment = new TackleAttachmentService(this.Monitor, key => helper.Translation.Get(key));
         ConfigValidationReport report = this.configManager.Load();
 
         this.Monitor.Log(
@@ -118,6 +120,7 @@ internal sealed class ModEntry : Mod
     private void OnUpdateTicked(object? sender, UpdateTickedEventArgs e)
     {
         this.baitAttachment!.UpdateCurrent(this.configManager!.Active);
+        this.tackleAttachment!.UpdateCurrent(this.configManager.Active);
         this.automationRuntime!.UpdateCurrent();
     }